Live Plotting: adds motor current plotting, debug
This commit adds some extra features for live-plotting data. It allows the user
to drag a few windows of the power usage (current * voltage) for each motor. It
also adds a new feature to DeltaStation called "debug mode", in which
pseudo-randomized data is generated and parsed by ds. This allows us to develop
features remotely without needing a serial input setup.