config: toml: add toml-based configuration
This is a large overhaul to the codebase. It currently disables many of
the graphs added to Window.cpp in favor of testing future graphs that
hinge on expression parsing alone. Some graphs will have to be
hard-coded (i.e. live GPS, strategy sending, graph switchboard, etc.)

Expression parsing is implemented in the next commit. For this one, it
mainly concerns automatic generation of buffers. In short, the
configuration of buffers as listed in a config.toml file will determine
the memory layout of messages coming from the car's telemetry subsystem.