config: expr: add expression parsing subsystem
Any algebraic combination of the real numbers and members of the
DeltaStation buffer configuration can now be represented as an object
that produces a number whenever values for the buffer config members are
provided. In short, we will eventually be able to graph live changes in
arbitrary expressions involving e.g. motor voltage times motor current,
etc.