expr: graph: enable simple custom live graphing
We can now do live graphing of custom-defined graphs. It also adds the
graphing files to the project-level CMakeLists.txt, as those were
removed in a previous commit to enable the project to build. It
implements a test or verifying the lexer and parser work. Further
revisions of these tools will use more advanced complier features (such
as regex, function parsing, a standard library with cos, sin, exp, etc.)